+++ .vscode/settings.json
@@ -2,6 +2,8 @@
   "rust-analyzer.cargo.features": "all",
   "rust-analyzer.linkedProjects": [
     "./Cargo.toml",
-    "./lib/Cargo.toml"
+    "./lib/Cargo.toml",
+    "./raiko-host/Cargo.toml",
+    "./raiko-host/Cargo.toml"
   ]
 }
