+++ lib/Cargo.toml
@@ -4,19 +4,19 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0"
-alloy-sol-types = { version = "0.4", optional = true }
+anyhow = { version = "1.0", default-features = false }
+alloy-sol-types = { version = "0.6", default-features = false, optional = true }
 ethers-core = { version = "2.0", features = ["optimism"] }
 hashbrown = { workspace = true }
-once_cell = "1.18"
+once_cell = { version = "1.18", default-features = false }
 revm = { workspace = true }
 ruint = { version = "1.10", default-features = false }
-serde = "1.0"
-thiserror = "1.0"
+serde = { version = "1.0", default-features = false, features = ["alloc"] }
+thiserror-no-std = "2.0.2"
 zeth-primitives = { path = "../primitives", features = ["revm"] }
-hex = "0.4.3"
-rlp = "0.5.2"
-tracing = "0.1"
+hex = { version = "0.4.3", optional = true}
+rlp = { version = "0.5.2", optional = true}
+tracing = { version = "0.1", optional = true}
 
 [target.'cfg(not(target_os = "zkvm"))'.dependencies]
 chrono = { version = "0.4", default-features = false }
@@ -33,6 +33,14 @@ bincode = "1.3"
 serde_with = "3.1"
 
 [features]
-taiko = ["zeth-primitives/taiko", "dep:alloy-sol-types"]
+default = ["std", "taiko"]
+std = []
+taiko = [
+    "zeth-primitives/taiko", 
+    "dep:alloy-sol-types",
+    "dep:hex",
+    "dep:rlp",
+    "dep:tracing"
+    ]
 pos = []
 server = []
