+++ lib/src/execution/ethereum.rs
@@ -11,8 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-use std::{fmt::Debug, mem::take};
+use alloc::format;
+use core::{fmt::Debug, mem::take};
 
 use anyhow::{anyhow, bail, Context};
 #[cfg(not(target_os = "zkvm"))]
@@ -160,9 +160,11 @@ impl TxExecStrategy<EthereumTxEssence> for EthTxExecStrategy {
             let trie_key = tx_no.to_rlp();
             tx_trie
                 .insert_rlp(&trie_key, tx)
+                .map_err(Into::<anyhow::Error>::into)
                 .context("failed to insert transaction")?;
             receipt_trie
                 .insert_rlp(&trie_key, receipt)
+                .map_err(Into::<anyhow::Error>::into)
                 .context("failed to insert receipt")?;
 
             // update account states
@@ -222,6 +224,7 @@ impl TxExecStrategy<EthereumTxEssence> for EthTxExecStrategy {
             // Add withdrawal to trie
             withdrawals_trie
                 .insert_rlp(&i.to_rlp(), withdrawal)
+                .map_err(Into::<anyhow::Error>::into)
                 .context("failed to insert withdrawal")?;
         }
