+++ lib/src/taiko/execute.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{fmt::Debug, mem::take, str::from_utf8};
-
+use core::{fmt::Debug, mem::take, str::from_utf8};
+use alloc::vec;
 use anyhow::{anyhow, bail, Context};
 #[cfg(all(not(target_os = "zkvm"), not(feature = "server")))]
 use log::debug;
@@ -282,6 +282,7 @@ impl TxExecStrategy<EthereumTxEssence> for TaikoTxExecStrategy {
             // Add withdrawal to trie
             withdrawals_trie
                 .insert_rlp(&i.to_rlp(), withdrawal)
+                .map_err(Into::<anyhow::Error>::into)
                 .with_context(|| "failed to insert withdrawal")?;
         }
