--- lib/src/taiko/host.rs
@@ -1,313 +1,0 @@
-use anyhow::Result;
-use ethers_core::types::{Block, Transaction as EthersTransaction, H160, H256, U256};
-use tracing::info;
-use zeth_primitives::{
-    ethers::{from_ethers_h160, from_ethers_h256, from_ethers_u256},
-    taiko::*,
-    transactions::ethereum::EthereumTxEssence,
-    withdrawal::Withdrawal,
-    Address, B256,
-};
-
-use crate::{
-    block_builder::{BlockBuilder, NetworkStrategyBundle},
-    consts::ChainSpec,
-    host::{
-        provider::{new_provider, BlockQuery, ProofQuery, ProposeQuery, Provider},
-        Init,
-    },
-    input::Input,
-    taiko::{precheck::rebuild_and_precheck_block, Layer},
-};
-
-#[derive(Debug)]
-pub struct TaikoExtra {
-    pub l1_hash: B256,
-    pub l1_height: u64,
-    pub l2_tx_list: Vec<u8>,
-    pub prover: Address,
-    pub graffiti: B256,
-    pub l1_signal_root: B256,
-    pub l2_signal_root: B256,
-    pub l2_withdrawals: Vec<Withdrawal>,
-    pub block_proposed: BlockProposed,
-    pub l1_next_block: Block<EthersTransaction>,
-    pub l2_fini_block: Block<EthersTransaction>,
-}
-
-#[allow(clippy::type_complexity)]
-fn fetch_data(
-    annotation: &str,
-    cache_path: Option<String>,
-    rpc_url: Option<String>,
-    block_no: u64,
-    signal_service: Address,
-    layer: Layer,
-) -> Result<(
-    Box<dyn Provider>,
-    Block<H256>,
-    Block<EthersTransaction>,
-    B256,
-    Input<EthereumTxEssence>,
-)> {
-    let mut provider = new_provider(cache_path, rpc_url)?;
-
-    let fini_query = BlockQuery { block_no };
-    match layer {
-        Layer::L1 => {}
-        Layer::L2 => {
-            provider.batch_get_partial_blocks(&fini_query)?;
-        }
-    }
-    // Fetch the initial block
-    let init_block = provider.get_partial_block(&BlockQuery {
-        block_no: block_no - 1,
-    })?;
-
-    info!(
-        "Initial {} block: {:?} ({:?})",
-        annotation,
-        init_block.number.unwrap(),
-        init_block.hash.unwrap()
-    );
-
-    // Fetch the finished block
-    let fini_block = provider.get_full_block(&fini_query)?;
-
-    info!(
-        "Final {} block number: {:?} ({:?})",
-        annotation,
-        fini_block.number.unwrap(),
-        fini_block.hash.unwrap()
-    );
-    info!("Transaction count: {:?}", fini_block.transactions.len());
-
-    // Get l2 signal root by signal service
-    let proof = provider.get_proof(&ProofQuery {
-        block_no,
-        address: H160::from_slice(signal_service.as_slice()),
-        indices: Default::default(),
-    })?;
-    let signal_root = from_ethers_h256(proof.storage_hash);
-
-    info!(
-        "Final {} signal root: {:?} ({:?})",
-        annotation,
-        fini_block.number.unwrap(),
-        signal_root,
-    );
-
-    // Create input
-    let input = Input {
-        beneficiary: fini_block.author.map(from_ethers_h160).unwrap_or_default(),
-        gas_limit: from_ethers_u256(fini_block.gas_limit),
-        timestamp: from_ethers_u256(fini_block.timestamp),
-        extra_data: fini_block.extra_data.0.clone().into(),
-        mix_hash: from_ethers_h256(fini_block.mix_hash.unwrap()),
-        transactions: fini_block
-            .transactions
-            .clone()
-            .into_iter()
-            .map(|tx| tx.try_into().unwrap())
-            .collect(),
-        withdrawals: fini_block
-            .withdrawals
-            .clone()
-            .unwrap_or_default()
-            .into_iter()
-            .map(|w| w.try_into().unwrap())
-            .collect(),
-        parent_state_trie: Default::default(),
-        parent_storage: Default::default(),
-        contracts: vec![],
-        parent_header: init_block.clone().try_into()?,
-        ancestor_headers: vec![],
-        base_fee_per_gas: from_ethers_u256(fini_block.base_fee_per_gas.unwrap_or_default()),
-    };
-
-    Ok((provider, init_block, fini_block, signal_root, input))
-}
-
-fn execute_data<N: NetworkStrategyBundle<TxEssence = EthereumTxEssence>>(
-    provider: Box<dyn Provider>,
-    chain_spec: ChainSpec,
-    init_block: Block<H256>,
-    input: Input<EthereumTxEssence>,
-    fini_block: Block<EthersTransaction>,
-) -> Result<Init<EthereumTxEssence>> {
-    // Create the provider DB
-    let provider_db =
-        crate::host::provider_db::ProviderDb::new(provider, init_block.number.unwrap().as_u64());
-    // Create the block builder, run the transactions and extract the DB
-    let mut builder = BlockBuilder::new(&chain_spec, input)
-        .with_db(provider_db)
-        .prepare_header::<N::HeaderPrepStrategy>()?
-        .execute_transactions::<N::TxExecStrategy>()?;
-    let provider_db = builder.mut_db().unwrap();
-
-    info!("Gathering inclusion proofs ...");
-
-    // Gather inclusion proofs for the initial and final state
-    let init_proofs = provider_db.get_initial_proofs()?;
-    let fini_proofs = provider_db.get_latest_proofs()?;
-
-    // Gather proofs for block history
-    let history_headers = provider_db.provider.batch_get_partial_blocks(&BlockQuery {
-        block_no: fini_block.number.unwrap().as_u64(),
-    })?;
-    // ancestors == history - current - parent
-    let ancestor_headers = if history_headers.len() > 2 {
-        history_headers
-            .into_iter()
-            .rev()
-            .skip(2)
-            .map(|header| {
-                header
-                    .try_into()
-                    .expect("Failed to convert ancestor headers")
-            })
-            .collect()
-    } else {
-        vec![]
-    };
-
-    info!("Saving provider cache ...");
-
-    // Save the provider cache
-    provider_db.get_provider().save()?;
-    info!("Provider-backed execution is Done!");
-    // assemble init
-    let transactions = fini_block
-        .transactions
-        .clone()
-        .into_iter()
-        .map(|tx| tx.try_into().unwrap())
-        .collect();
-    let withdrawals = fini_block
-        .withdrawals
-        .clone()
-        .unwrap_or_default()
-        .into_iter()
-        .map(|w| w.try_into().unwrap())
-        .collect();
-
-    let init = Init {
-        db: provider_db.get_initial_db().clone(),
-        init_block: init_block.try_into()?,
-        init_proofs,
-        fini_block: fini_block.try_into()?,
-        fini_transactions: transactions,
-        fini_withdrawals: withdrawals,
-        fini_proofs,
-        ancestor_headers,
-    };
-    Ok(init)
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn get_taiko_initial_data<N: NetworkStrategyBundle<TxEssence = EthereumTxEssence>>(
-    l1_cache_path: Option<String>,
-    _l1_chain_spec: ChainSpec,
-    l1_rpc_url: Option<String>,
-    prover: Address,
-    l2_cache_path: Option<String>,
-    l2_chain_spec: ChainSpec,
-    l2_rpc_url: Option<String>,
-    l2_block_no: u64,
-    graffiti: B256,
-) -> Result<(Init<EthereumTxEssence>, TaikoExtra)> {
-    let (l2_provider, l2_init_block, mut l2_fini_block, l2_signal_root, l2_input) = fetch_data(
-        "L2",
-        l2_cache_path,
-        l2_rpc_url,
-        l2_block_no,
-        l2_chain_spec.l2_signal_service.unwrap(),
-        Layer::L2,
-    )?;
-    // Get anchor call parameters
-    let anchorCall {
-        l1Hash: anchor_l1_hash,
-        l1SignalRoot: anchor_l1_signal_root,
-        l1Height: l1_block_no,
-        parentGasUsed: l2_parent_gas_used,
-    } = decode_anchor_call_args(&l2_fini_block.transactions[0].input)?;
-
-    let (mut l1_provider, _l1_init_block, l1_fini_block, l1_signal_root, _l1_input) = fetch_data(
-        "L1",
-        l1_cache_path,
-        l1_rpc_url,
-        l1_block_no,
-        l2_chain_spec.l1_signal_service.unwrap(),
-        Layer::L1,
-    )?;
-
-    let (propose_tx, block_metadata) = l1_provider.get_propose(&ProposeQuery {
-        l1_contract: H160::from_slice(l2_chain_spec.l1_contract.unwrap().as_slice()),
-        l1_block_no: l1_block_no + 1,
-        l2_block_no,
-    })?;
-
-    let l1_next_block = l1_provider.get_full_block(&BlockQuery {
-        block_no: l1_block_no + 1,
-    })?;
-
-    // save l1 data
-    l1_provider.save()?;
-
-    let proposeBlockCall {
-        params: _,
-        txList: l2_tx_list,
-    } = decode_propose_block_call_args(&propose_tx.input)?;
-
-    // 1. check l2 parent gas used
-    if l2_init_block.gas_used != U256::from(l2_parent_gas_used) {
-        return Err(anyhow::anyhow!(
-            "parent gas used mismatch, expect: {}, got: {}",
-            l2_init_block.gas_used,
-            l2_parent_gas_used
-        ));
-    }
-    // 2. check l1 signal root
-    if anchor_l1_signal_root != l1_signal_root {
-        return Err(anyhow::anyhow!(
-            "l1 signal root mismatch, expect: {}, got: {}",
-            anchor_l1_signal_root,
-            l1_signal_root
-        ));
-    }
-    // 3. check l1 block hash
-    if Some(anchor_l1_hash) != l1_fini_block.hash.map(from_ethers_h256) {
-        return Err(anyhow::anyhow!(
-            "l1 block hash mismatch, expect: {}, got: {:?}",
-            anchor_l1_hash,
-            l1_fini_block.hash
-        ));
-    }
-
-    let extra = TaikoExtra {
-        l1_hash: anchor_l1_hash,
-        l1_height: l1_block_no,
-        l2_tx_list,
-        prover,
-        graffiti,
-        l1_signal_root,
-        l2_signal_root,
-        l2_withdrawals: l2_input.withdrawals.clone(),
-        block_proposed: block_metadata,
-        l1_next_block,
-        l2_fini_block: l2_fini_block.clone(),
-    };
-
-    // rebuild transaction list by tx_list from l1 contract
-    rebuild_and_precheck_block(&l2_chain_spec, &mut l2_fini_block, &extra)?;
-
-    // execute transactions and get states
-    let init = execute_data::<N>(
-        l2_provider,
-        l2_chain_spec,
-        l2_init_block,
-        l2_input,
-        l2_fini_block,
-    )?;
-    Ok((init, extra))
-}
