+++ lib/src/taiko/utils.rs
@@ -1,3 +1,4 @@
+use alloc::vec::Vec;
 use rlp::{Decodable, DecoderError, Rlp};
 
 pub fn rlp_decode_list<T>(bytes: &[u8]) -> Result<Vec<T>, DecoderError>
