+++ primitives/src/taiko/proposal.rs
@@ -16,3 +16,4 @@ pub fn decode_propose_block_call_args(data: &[u8]) -> Result<proposeBlockCall> {
         .with_context(|| "failed to decode propose block call")?;
     Ok(propose_block_call)
 }
+
