+++ raiko-host/Cargo.toml
@@ -3,6 +3,10 @@ name = "raiko-host"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+name = "util"
+path = "src/util/lib.rs"
+
 [dependencies]
 anyhow = "1.0"
 bincode = "1.3.3"
@@ -34,3 +38,7 @@ once_cell = "1.8.0"
 [dev-dependencies]
 assert_cmd = "2.0"
 rstest = "0.18"
+
+[features]
+default = ["taiko"]
+taiko = []
